Create ApplicantLog model.
