+++ huntflow_api_client/entities/__init__.py
@@ -1,6 +1,7 @@
 from huntflow_api_client.entities.account_offers import AccountOffer
 from huntflow_api_client.entities.account_vacancy_request import AccountVacancyRequest
 from huntflow_api_client.entities.accounts import Account
+from huntflow_api_client.entities.applicant_logs import ApplicantLog
 from huntflow_api_client.entities.applicant_offers import ApplicantOffer
 from huntflow_api_client.entities.applicant_on_vacancy import ApplicantOnVacancy
 from huntflow_api_client.entities.applicant_on_vacancy_status import ApplicantOnVacancyStatus
@@ -27,6 +28,7 @@ __all__ = (
     "AccountTag",
     "AccountVacancyRequest",
     "Applicant",
+    "ApplicantLog",
     "ApplicantOffer",
     "ApplicantOnVacancy",
     "ApplicantOnVacancyStatus",
